+++ GanJiJakSalPUBG/Assets/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 392938131b17c5a49a495e0e2b58988f
+guid: 68ad6fc77a403994db4e920853fe974d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
